// src/models/Github.ts
export interface GithubPayload {
  action: string;
  number: number;
  "pull_request": PullRequest;
  sender: Sender;
  repository: Repository;
}

export interface PullRequest {
  url: string;
  id: number;
  "node_id": string;
  "html_url": string;
  "diff_url": string;
  "patch_url": string;
  "issue_url": string;
  number: number;
  state: string;
  locked: boolean;
  title: string;
  user: Sender;
  body: string;
  "created_at": string;
  "updated_at": string;
  "closed_at": null;
  "merged_at": null;
  "merge_commit_sha": null;
  assignee: null;
  assignees: any[];
  "requested_reviewers": Sender[];
  "requested_teams": any[];
  labels: Label[];
  milestone: null;
  draft: boolean;
  "commits_url": string;
  "review_comments_url": string;
  "review_comment_url": string;
  "comments_url": string;
  "statuses_url": string;
  "author_association": string;
  "auto_merge": null;
  "active_lock_reason": null;
  merged: boolean;
  mergeable: null;
  rebaseable: null;
  "mergeable_state": string;
  "merged_by": null;
  comments: number;
  "review_comments": number;
  "maintainer_can_modify": boolean;
  commits: number;
  additions: number;
  deletions: number;
  "changed_files": number;
}

export interface Repository {
  id: number;
  name: string;
  "html_url": string;
}

// Generated by https://quicktype.io

export interface Label {
  name: string;
  color: string;
  description: string;
}

export interface Sender {
  login: string;
  id: number;
  "node_id": string;
  "avatar_url": string;
  url: string;
  type: string;
}